requirements.txt
openai>=1.0.0
google-genai>=1.0.0
pypdf>=3.0.0
nbformat>=5.9.0
nbconvert>=7.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
ipykernel>=6.25.0
